+++ app/src/main/AndroidManifest.xml
@@ -151,7 +151,7 @@
             android:screenOrientation="portrait" />
 
         <activity
-            android:name="com.going.presentation.checkfriends.CheckFriendsActivity"
+            android:name="com.going.presentation.todo.ourtodo.checkfriends.CheckFriendsActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
