+++ presentation/src/main/java/com/going/presentation/entertrip/createtrip/preference/EnterPreferenceActivity.kt
@@ -22,8 +22,8 @@ import com.going.presentation.entertrip.invitetrip.invitecode.CreateTripActivity
 import com.going.presentation.entertrip.invitetrip.invitecode.CreateTripActivity.Companion.START_DAY
 import com.going.presentation.entertrip.invitetrip.invitecode.CreateTripActivity.Companion.START_MONTH
 import com.going.presentation.entertrip.invitetrip.invitecode.CreateTripActivity.Companion.START_YEAR
-import com.going.presentation.preferencetag.PreferenceTagAdapter
-import com.going.presentation.preferencetag.PreferenceTagDecoration
+import com.going.presentation.entertrip.preferencetag.PreferenceTagAdapter
+import com.going.presentation.entertrip.preferencetag.PreferenceTagDecoration
 import com.going.ui.base.BaseActivity
 import com.going.ui.extension.UiState
 import com.going.ui.extension.setOnSingleClickListener
