+++ presentation/src/main/java/com/going/presentation/entertrip/invitetrip/preference/FinishPreferenceActivity.kt
@@ -11,8 +11,8 @@ import com.going.presentation.R
 import com.going.presentation.dashboard.DashBoardActivity
 import com.going.presentation.databinding.ActivityFinishPreferenceBinding
 import com.going.presentation.entertrip.invitetrip.finish.InviteFinishActivity
-import com.going.presentation.preferencetag.PreferenceTagAdapter
-import com.going.presentation.preferencetag.PreferenceTagDecoration
+import com.going.presentation.entertrip.preferencetag.PreferenceTagAdapter
+import com.going.presentation.entertrip.preferencetag.PreferenceTagDecoration
 import com.going.ui.base.BaseActivity
 import com.going.ui.extension.UiState
 import com.going.ui.extension.setOnSingleClickListener
